+++ src/hls/stream.cpp
@@ -73,7 +73,7 @@ void Stream::go_to_next_segment() {
 
 uint64_t Stream::get_total_duration() {
   std::lock_guard<std::mutex> lock(data_mutex);
-  uint32_t total_time = 0;
+  uint64_t total_time = 0;
   for(auto it = segments.begin(); it != segments.end(); ++it) {
     total_time += it->duration;
   }
